refactor: 파일 이름 수정(#111)
- 줄임말을 최대한 사용하지 않는 방향으로 파일 이름을 수정하였습니다.
